feat: create separated Test and Game scenes

+++ Assets/Scripts/Utils.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Utils : MonoBehaviour
 {
-    static public float ConvertBlockScaleToPosition(int length)
+    static public float ConvertBlockScaleToPosition(float length)
 	{
 		return length / 100f;
 	}
